app/requirements.txt
numpy
streamlit
pandas
matplotlib.pyplot
scikit-learn